vtrent OS

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: vtrentos <vtrentos@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/01 10:31:30 by Vtrentos      #+#    #+#                 */
-/*   Updated: 2020/12/10 12:38:36 by Vtrentos      ########   odam.nl         */
+/*   Updated: 2021/01/05 19:53:18 by Vtrentos      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned long int i;
+	size_t i;
 
 	i = 0;
 	while (*s != '\0')
